+++ NEWS.md
@@ -1,4 +1,16 @@
-# V0.1.0
+# pairplot 0.1.1
+
+## Changes
+
+* Remove pairgrid unused arguments.
+* Add possible knowledge of surrounding subplots for diagonal plots.
+
+## Bug Fixes
+
+* **Important**: Fixed min-max scaling error in `pair_geom_histogram`.
+
+
+# pairplot 0.1.0
 
 * The layout system relies on [patchwork](https://patchwork.data-imaginist.com/index.html)
 * Now accessible on GitHub.